+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.1.0] - 2026-05-20
 
 ### Added
 
@@ -29,5 +29,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial stable version.
 
+[1.1.0]: https://github.com/ozontech/testo/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/ozontech/testo/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/ozontech/testo/releases/tag/v1.0.0